ehancement based on meeting result 15 march 2022
- add date order when create order
- add sync product
- add api to get product detail when scan
- add feature to edit order for unsynced order

+++ app/Http/Controllers/SyncController.php
@@ -5,10 +5,15 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Jobs\ProcessOrder;
 use App\Models\Order;
+use Auth;
 
 class SyncController extends Controller
 {
     public function index(){
+        if (Auth::user()->cannot('sync', Auth::user())) {
+            return abort(403);
+        }
+
         $data['orders'] = Order::latest()->paginate(10);
         return view('sync',$data);
     }
@@ -21,14 +26,19 @@ class SyncController extends Controller
         
         ProcessOrder::dispatchSync($order);
 
-        return redirect()->route('sync');
+        return back();
     }
 
     public function delete($order_no)
     {
-        $order = Order::where('order_no',$order_no)->first();
-        $order->lines()->delete();
-        $order->delete();
-        return redirect()->route('sync');
+        $order = Order::whereNull('c_order_id')->where('order_no',$order_no)->first();
+
+        if(Auth::user()->is_admin || Auth::id()==$order->user_id){
+            $order->lines()->delete();
+            $order->delete();
+            return back()->with('message','Order has been deleted');
+        }
+
+        return abort(403);
     }
 }
